# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Dispatch controller API
  version: 0.1.0
  description: Project-scoped application deployment and operations API.
servers:
  - url: /api/v1
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /audit:
    get:
      operationId: listAuditEvents
      description: Requires authentication. Owners see controller events; members see events only in projects they
        can view. Returns at most 100 events.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditEvent'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      parameters:
      - name: projectId
        in: query
        schema:
          type: string
      - name: appId
        in: query
        schema:
          type: string
      - name: actorId
        in: query
        schema:
          type: string
      - name: action
        in: query
        schema:
          type: string
      - name: before
        in: query
        schema:
          type: string
  /apps/{id}/owner:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      operationId: getApplicationOwner
      description: Requires project.view. Responsible person/team metadata and display name. Ownership does not grant
        permissions.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationOwner'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
    put:
      operationId: setApplicationOwner
      description: Requires project.configure. Record responsible user/team, or clear principalId to remove the label.
        Does not modify access grants.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationOwner'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                principalType:
                  type: string
                  enum:
                  - user
                  - team
                principalId:
                  type: string
  /projects/{id}/owner-candidates:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      operationId: listApplicationOwnerCandidates
      description: Requires project.configure. Lists eligible user/team metadata for ownership selection within the
        project.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      type: object
                  teams:
                    type: array
                    items:
                      type: object
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
  /projects/{id}/retention:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      operationId: getProjectRetention
      description: Requires project.manage. Read an explicit history policy; no automatic cleanup is enabled.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetentionPolicy'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
    put:
      operationId: saveProjectRetention
      description: Requires project.manage. Store a retention policy. Cleanup is a separate explicit operation.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetentionPolicy'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetentionPolicy'
  /projects/{id}/retention/preview:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    post:
      operationId: previewProjectRetention
      description: Requires project.manage. Count eligible historical logs/runs, preserving current, successful, baseline,
        and workflow dependencies.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetentionResult'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /projects/{id}/retention/apply:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    post:
      operationId: applyProjectRetention
      description: Requires project.manage. Reevaluate eligibility and delete only unprotected data in this project.
        This operation is irreversible.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetentionResult'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - confirm
              properties:
                confirm:
                  type: string
                  description: Must equal the project ID.
  /operations/backups:
    get:
      operationId: listControllerBackups
      description: Controller owner only. Metadata and restore-verification outcomes, never backup paths, contents,
        or credentials.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured:
                    type: boolean
                  backups:
                    type: array
                    items:
                      $ref: '#/components/schemas/BackupRecord'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
    post:
      operationId: createControllerBackup
      description: Controller owner only. Create a consistent controller database and encrypted credential key backup
        in protected controller storage.
      responses:
        '201':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupRecord'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /operations/backups/{id}/verify:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    post:
      operationId: verifyControllerBackup
      description: Controller owner only. Verify a backup in disposable storage/database. Does not replace the running
        controller.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackupRecord'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /services/{id}/impact:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      operationId: getServiceImpact
      description: Requires project.view. List dependent applications, environments, applied revisions and redeployment
        availability.
      responses:
        '200':
          description: Successful result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceImpact'
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /services/{id}/redeploy:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    post:
      operationId: redeployServiceConsumers
      description: Requires project.configure and deployment.run. Explicitly redeploy selected consumers with the
        last successful source revision and current service inputs. Repository-managed consumers use workflow promotion.
      responses:
        '202':
          description: Successful result.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    appId:
                      type: string
                    deployment:
                      $ref: '#/components/schemas/Deployment'
                    error:
                      type: string
        '400':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - revision
              - appIds
              - confirm
              properties:
                revision:
                  type: integer
                appIds:
                  type: array
                  items:
                    type: string
                confirm:
                  type: boolean
                  const: true
...
  /services/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      operationId: getService
      responses:
        "200":
          description: Connection metadata. Sensitive values are never returned.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
    put:
      operationId: updateService
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServiceInput" }
      responses:
        "200": { description: Updated. Connection edits increment revision and clear the check result. }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteService
      responses:
        "204": { description: Registration removed. The external dependency is untouched. }
        "409": { $ref: "#/components/responses/Problem" }
...
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectID"
    put:
      operationId: updateProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "200": { description: Project updated. }
        "404": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteProject
      responses:
        "204": { description: Project deleted. }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
  /apps/{appId}:
    parameters:
      - $ref: "#/components/parameters/AppID"
    delete:
      operationId: deleteApp
      description: Cleans deployed resources, then transactionally deletes the application and its deployment history.
      responses:
        "204": { description: Application and dependent history deleted. }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
  /deployments/{deploymentId}:
    get:
      operationId: getDeployment
      parameters:
        - $ref: "#/components/parameters/DeploymentID"
      responses:
        "200":
          description: Deployment evidence.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Deployment" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectID:
      in: path
      name: projectId
      required: true
      schema: { type: string }
...
    AppID:
      in: path
      name: appId
      required: true
      schema: { type: string }
    DeploymentID:
      in: path
      name: deploymentId
      required: true
      schema: { type: string }
...
  responses:
...
    Problem:
      description: RFC 9457-style problem response.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ApplicationOwner:
      type: object
      properties:
        appId:
          type: string
        principalType:
          type: string
          enum:
          - user
          - team
          - ''
        principalId:
          type: string
        displayName:
          type: string
          readOnly: true
        updatedAt:
          type: string
          format: date-time
    AuditEvent:
      type: object
      properties:
        id:
          type: string
        actorId:
          type: string
        actorName:
          type: string
        impersonatorId:
          type: string
        projectId:
          type: string
        appId:
          type: string
        action:
          type: string
        resourceId:
          type: string
        outcome:
          type: string
        createdAt:
          type: string
          format: date-time
...
    RetentionPolicy:
      type: object
      required:
      - logDays
      - runDays
      - keepRuns
      properties:
        projectId:
          type: string
        logDays:
          type: integer
          minimum: 1
          maximum: 36500
        runDays:
          type: integer
          minimum: 1
          maximum: 36500
        keepRuns:
          type: integer
          minimum: 5
          maximum: 10000
    RetentionResult:
      type: object
      properties:
        logs:
          type: integer
        runs:
          type: integer
        protectedRuns:
          type: integer
        applied:
          type: boolean
    BackupRecord:
      type: object
      properties:
        id:
          type: string
        engine:
          type: string
        state:
          type: string
        bytes:
          type: integer
        createdAt:
          type: string
          format: date-time
        verifiedAt:
          type: string
          format: date-time
        message:
          type: string
    ServiceImpact:
      type: object
      properties:
        serviceId:
          type: string
        revision:
          type: integer
        consumers:
          type: array
          items:
            type: object
            properties:
              appId:
                type: string
              appName:
                type: string
              alias:
                type: string
              appliedRevision:
                type: integer
              redeploymentRequired:
                type: boolean
              targetName:
                type: string
              environment:
                type: string
              deploymentId:
                type: string
              canRedeploy:
                type: boolean
              reason:
                type: string
...
    ServiceInput:
      type: object
      description: Project, name and type are required at creation and immutable thereafter. Omitted fields are preserved. Only the controller owner can attach global secret references.
      properties:
        projectId: { type: string }
        name: { type: string, pattern: "^[a-z0-9][a-z0-9.-]{0,62}$" }
        type: { type: string, enum: [postgresql, generic] }
        description: { type: string }
        revision: { type: integer, description: Expected current revision for optimistic concurrency. }
        connectionUrl: { type: string, writeOnly: true, description: PostgreSQL URL input instead of individual fields. }
        probeHost: { type: string }
        probePort: { type: integer }
        fields:
          type: object
          additionalProperties:
            type: object
            properties:
              value: { type: string, writeOnly: true, maxLength: 65536 }
              sensitive: { type: boolean }
              secretRef: { type: string }
              remove: { type: boolean }
    Service:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        name: { type: string }
        type: { type: string, enum: [postgresql, generic] }
        description: { type: string }
        revision: { type: integer }
        fields:
          type: object
          additionalProperties:
            type: object
            properties:
              value: { type: string, description: Ordinary fields only. }
              sensitive: { type: boolean }
              configured: { type: boolean }
              secretRef: { type: string, description: Visible only to the controller owner. }
        availableFields: { type: array, items: { type: string } }
        check: { $ref: "#/components/schemas/ServiceCheck" }
        consumers:
          type: array
          items:
            type: object
            properties:
              appId: { type: string }
              appName: { type: string }
              alias: { type: string }
              appliedRevision: { type: integer }
              redeploymentRequired: { type: boolean }
    ServiceCheck:
      type: object
      properties:
        state: { type: string, enum: [succeeded, failed, untested] }
        message: { type: string }
        location: { type: string, const: Dispatch controller }
        checkedAt: { type: string, format: date-time }
        durationMs: { type: integer }
...
    Deployment:
      type: object
      required: [id, appId, commitSha, specDigest, state, message, createdAt]
      properties:
        id: { type: string }
        appId: { type: string }
        commitSha: { type: string }
        specDigest: { type: string }
        state:
          type: string
          enum: [queued, fetching, building, starting, checking, routing, succeeded, failed, cancelled]
        message: { type: string }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
    Problem:
      type: object
      required: [status, title, detail]
      properties:
        status: { type: integer }
        title: { type: string }
        detail: { type: string }